+++ datacrawler/transform/transforme_le_bloc_budget_et_finances_des_établissements_médico_sociaux/transforme_et_fusionne_les_données_budgétaires_avec_les_taux_des_établissements_errd_test.py
@@ -3,13 +3,11 @@ from numpy import NaN
 
 from datacrawler.test_helpers import (
     NUMÉRO_FINESS_ÉTABLISSEMENT,
-    csv_ann_errd_ej_builder,
-    csv_ann_errd_ej_et_budget_et_finances_builder,
-    csv_ann_per_errd_eprd_builder,
     helios_ann_errd_ej_budget_et_finances_builder,
     helios_ann_errd_ej_et_budget_et_finances_builder,
     mocked_logger,
 )
+from datacrawler.test_helpers.diamant_builder import csv_ann_errd_ej_builder, csv_ann_errd_ej_et_budget_et_finances_builder, csv_ann_per_errd_eprd_builder
 from datacrawler.transform.transforme_le_bloc_budget_et_finances_des_établissements_médico_sociaux.transforme_et_fusionne_les_données_budgétaires_avec_les_taux_des_établissements_errd import (
     transforme_et_fusionne_les_données_budgétaires_errd_avec_les_taux_errd,
 )
